Refactor new components for footer, navbar, hero, about, and testimonials
