+++ index.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
+  <head style="background-color: #121212">
     <meta charset="UTF-8" />
     <link
       rel="stylesheet"
@@ -13,7 +13,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Jaakko Lipponen</title>
   </head>
-  <body>
+  <body style="background-color: #121212">
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
   </body>
